Finding the best train for arrival time api method added

+++ api/models/train.go
@@ -8,5 +8,5 @@ type Train struct {
 	Id       bson.ObjectId `json:"id" bson:"_id"`
 	Number   string        `form:"Number" json:"Number" bson:"Number"`
 	Name     string        `form:"Name" json:"Name"  bson:"Name"`
-	Schedule map[string]int64    `form:"Schedule" json:"Schedule" bson:"Schedule"`
+	Schedule map[string]int    `form:"Schedule" json:"Schedule" bson:"Schedule"`
 }
